fix(fs.service.js): `fs.service` was added to manage file-system indexing more efficient
`fs.service` was added to manage file-system indexing more efficient. It also fixes
"twice-event-report" problem caused by `git` command (#21).
This service (that is a singleton) can cache the result of requests and so improves the performance.
It exposes an extra layer on top of `readDirectory` that consumers (`buildConfig.js`, `refresh.js`,
and `watch.js` for now) should use (instead of direct access to `readDirectory`).

fix #21